Create dedicated app for editor #63
Manually migration of reverse track functionality is done
