+++ TrackEditorWeb/settings.py
@@ -32,6 +32,7 @@ ALLOWED_HOSTS = []
 
 INSTALLED_APPS = [
     'TrackApp',
+    'editor',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
